FIX: fix
미니게임 4 종료 후 이미지·GDI+ 리소스 및 릴스 효과음 미해제
미니게임 4 결과 문자열의 메모리 경계 경고
미니게임 1 폰트 경로 오류
분기 진입 실패 시 선택 횟수가 잘못 증가
BOOL/bool, 종료 코드 타입 경고

+++ No_Save_Love/No_Save_Love/main.cpp
@@ -45,7 +45,7 @@ int WINAPI WinMain(HINSTANCE hInstance, HINSTANCE hPrevinstance, LPSTR lpszCmdPa
         TranslateMessage(&Message);
         DispatchMessage(&Message);
     }
-    return Message.wParam;
+    return static_cast<int>(Message.wParam);
 }
 
 #include "GameManager.h"
